extract ObjectDetectionHelper.kt to separate class
